Move options away from redundant landUseLookup.ts
